+++ src/App.tsx
@@ -1,7 +1,6 @@
 import { createSignal, Show, onMount, createMemo } from "solid-js";
 import "./App.css";
 import Header from "./components/Header.tsx";
-import TrayNotificationToast from "./components/TrayNotificationToast.tsx";
 import SearchPage from "./pages/SearchPage.tsx";
 import InstalledPage from "./pages/InstalledPage.tsx";
 import { View } from "./types/scoop.ts";
@@ -237,7 +236,6 @@ function App() {
                     </div>
                 </div>
             </Show>
-            <TrayNotificationToast />
         </>
     );
 }