+++ RELEASE_NOTES.md
@@ -6,8 +6,14 @@
 
 #### ✨ Improvements
 
-* **NONE**
-
+* **Removed Update Detection native dialog popup**
+    - Now only shows in the banner update notification and on manual check for updates
+    - Removed unnecessary Notification toast inside Rscoop when closing the window
+*  **Tray workflow fixes**
+    - Displays more information
+    - Native dialog
+    - Rust parsing Scoop Apps using LNK Crate for efficiency and no shell dependancy
+*  **minor text improvements**
 #### 🐛 Bug Fixes
 
 * **Fixed being able to open multiple instances**